refactor(relay): general refactoring
